pyproject.toml
[project]
name = "azure-terminal-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.6.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
